chatbot/requirements.txt
annotated-types==0.7.0
archspec==0.2.3
argcomplete==3.5.1
asgiref==3.8.1
asyncio==3.4.3
autokey==0.96.0
Beaker==1.12.1
beautifulsoup4==4.12.3
blinker==1.9.0
blivet==3.10.0
blivet-gui==2.6.0
boltons==23.1.1
boto3==1.35.69
botocore==1.35.69
Brlapi==0.8.5
cachetools==5.5.2
certifi==2023.5.7
cffi==1.16.0
channels==4.2.0
charset-normalizer==3.3.2
click==8.1.7
conda==24.4.0
conda-libmamba-solver==23.11.1
conda-package-handling==2.2.0
conda_package_streaming==0.9.0
cryptography==41.0.7
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
distlib==0.3.9
distro==1.9.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
dnf==4.21.1
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
fros==1.1
frozendict==2.3.10
google-auth==2.38.0
google-genai==1.2.0
gTTS==2.5.3
guake==3.7.0
humanize==3.13.1
i3ipc==2.2.1
idna==3.7
iso639==0.1.4
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.3
langtable==0.0.68
libcomps==0.1.20
libdnf==0.73.3
libmambapy==1.5.8
louis==3.28.0
lxml==5.1.0
Mako==1.2.3
MarkupSafe==2.1.3
menuinst==2.0.0
meson==1.4.1
msgpack-python==0.5.6
msgpack-rpc-python==0.4.1
netifaces==0.11.0
nftables==0.1
numpy==2.2.2
nwg-panel==0.9.56
olefile==0.47
packaging==23.2
Paste==3.7.1
pbr==5.11.1
pexpect==4.9.0
pid==2.2.3
pillow==10.3.0
platformdirs==3.11.0
pluggy==1.3.0
ply==3.11
productmd==1.41
psutil==5.9.8
ptyprocess==0.7.0
pwquality==1.4.5
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.25.1
pycosat==0.6.6
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pydantic==2.10.6
pydantic_core==2.27.2
pyenchant==3.2.2
pygdbmi==0.11.0.0
PyGObject==3.48.2
pyinotify==0.9.6
pykickstart==3.52
pyOpenSSL==23.2.0
pyparted==3.13.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-meh==0.51
python-pam==2.0.2
python-xlib==0.33
pyudev==0.24.1
pyxdg==0.27
PyYAML==6.0.1
regex==2024.9.11
requests==2.31.0
requests-file==2.0.0
requests-ftp==0.3.1
rpm==4.19.1.1
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.7
s3transfer==0.10.4
selinux @ file:///builddir/build/BUILD/libselinux-3.7/src
sentry-sdk==2.17.0
sepolicy @ file:///builddir/build/BUILD/selinux-3.7/python/sepolicy
setools==4.5.1
setroubleshoot @ file:///builddir/build/BUILD/setroubleshoot-3.3.35/src
setuptools==69.0.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sos==4.7.2
soupsieve==2.5
sqlparse==0.5.3
systemd-python==235
Tempita==0.5.2
tornado==4.5.3
tqdm==4.66.5
typing_extensions==4.12.2
ufw==0.35
urllib3==1.26.20
uuid==1.30
virtualenv==20.29.1
websockets==14.2
Werkzeug==3.1.3
xkbregistry==0.3
zstandard==0.22.0
